+++ tools/tasks/browserSync.js
@@ -1,0 +1,1 @@
+module.exports = require('browser-sync').create();
