+++ tools/tasks/clean.js
@@ -1,0 +1,6 @@
+const shell = require('shelljs');
+
+module.exports = function clean() {
+    shell.rm('-rf', 'build');
+    shell.rm('-f', 'index.html');
+};
