+++ tools/tasks/isDev.js
@@ -1,0 +1,3 @@
+module.exports = function isDev() {
+    return process.env.NODE_ENV !== 'production';
+};
